fix bug in test due to NameError
In test test_pid_already_locked_multi_process a second process is being
used to test PidFile failing on a second lock for the same resource.

However since the exit code used checked for any failure the tests was
falsely succeeding as there also was a NameError present.

Fixed this by checking for a custom exit code and of course fixing the
NameError :-)